--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,5 @@
+module github.com/notnil/chess
+
+go 1.14
+
+require golang.org/x/text v0.3.2 // indirect