… go.sum
… zobrist_test.go