Version with [Zobrist hashing](https://en.wikipedia.org/wiki/Zobrist_hashing) ([original paper](https://research.cs.wisc.edu/techreports/1970/TR88.pdf)) instead of MD5.

I also have memoized the hash calculation.

```
$> benchstat bench_old.txt bench_new_v2.txt
name            old time/op    new time/op    delta
PositionHash-8    14.7µs ± 3%    13.2µs ± 3%  -10.32%  (p=0.000 n=10+9)

name            old alloc/op   new alloc/op   delta
PositionHash-8    9.63kB ± 1%    8.31kB ± 1%  -13.69%  (p=0.000 n=10+10)

name            old allocs/op  new allocs/op  delta
PositionHash-8       103 ± 0%        80 ± 0%  -22.33%  (p=0.000 n=10+10)
```

The difference between the 2 hashing methods will improve in the future with some changes like storing Castling rights as bitmaps and not as strings. One example can be found [here](https://github.com/dylhunn/dragontoothmg/blob/b0146de1e275d419ae46e1f36d3cec255d5a2a7b/types.go#L37,L129)

To search for collision due to bugs in my implementation I have also created a [tool](tools/zobroist_collisions/main.go) to compute hashes.