@notnil 2021-07-30
I would rather not have the collision tester as a separate main package.  I definitely get the importance of it, but would rather have a test case hidden behind a flag.  It's also not clear how the "mustParseFEN" statements were chosen.  I would use a list of games or random moves (with seed) test for collisions.