@notnil 2022-08-08
I still would eventually like to include this, but I believe it would be best to recycle ideas here into a separate PR.